Updating Model Controller and View
I updated these three classes to make the jump and fire animation cleaner. It interacts with the running smoothly. You stand still as you fire(otherwise you slide across the screen while orc is standing still which looks goofy). Created 'tap' functionality, this means when you tap the space bar or f key, the animation plays and you don't have to hold that key down. Can hold down fire key or jump key for looping of either functionality that looks how it should.

Still a pause at the first movement and sometimes lags when changing directions frequently. Overall lag is down I would say but more can be done to improve performance. Possibly loading image arrays as little as possible or doing so only once but for all(The latter might be slow as well but not sure)